# Optimization: 9~15% improvement in `KinesisDataFetcher` wall-time after