*Issue #, if available:*
n/a

*Description of changes:*
Optimization: 9~15% improvement in `KinesisDataFetcher` wall-time after converting `AWSExceptionManger` to a static variable.

Timed using multiple loops of 10k iterations/loop. Observed time decreased from `(p50=267ms, p100=398ms)` to `(p50=245ms, p100=307ms)`.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.